Add Main java file to execute counting CountingDictionary
This class will serve as the driver class for the entire project.
As the nature of this assignment was to turn in code to be reviewed, it
was not meant to be used by other people. Thus, adding this file will
allow others to interact with the code once they download it.